+++ sun8min-capital/capital-provider/src/main/java/com/sun8min/capital/provider/CapitalServiceImpl.java
@@ -6,6 +6,7 @@ import com.sun8min.capital.entity.Capital;
 import org.apache.dubbo.config.annotation.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service(version = "${service.version}")
@@ -34,7 +35,7 @@ public class CapitalServiceImpl implements CapitalService {
         return capitalDao.updateByPrimaryKey(record);
     }
 
-    public Long findAmountByUserId(Long userId) {
+    public BigDecimal findAmountByUserId(Long userId) {
         return capitalDao.findByUserId(userId).getCapitalAmount();
     }
 }
