+++ sun8min-redpacket/redpacket-entity/src/main/java/com.sun8min.redpacket.entity/RedpacketTradeOrder.java
@@ -1,17 +1,17 @@
 package com.sun8min.redpacket.entity;
 
+import java.io.Serializable;
+import java.math.BigDecimal;
+import java.util.Date;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-import java.util.Date;
 
 /**
  * 用户红包交易表
  *
  * @author sun8min
- * @date 2019-03-27 00:55:43
+ * @date 2019-03-28 13:26:03
  */
 @Data
 @NoArgsConstructor
@@ -33,9 +33,9 @@ public class RedpacketTradeOrder implements Serializable {
     private Long toUserId;
 
     /**
-     * 交易金额合计（精确到万分之一）
+     * 交易金额合计（精确到分）
      */
-    private Long redpacketTradeAmount;
+    private BigDecimal redpacketTradeAmount;
 
     /**
      * 订单交易号
